Some work on the Terminal
Simulator won't spam the input of the ACIA anymore, so you don't lose
characters!
Added a cursor! A very basic one (swaps the bg/fg)
Cleaned up some old comments
Fixed ANSI J and K (was putting the opcode in the first slot, which was
promptly deleted immediately)
Reinstated the Beep on boot, it's now in a location where it should 100%
work (wrapped in a try catch for safety reasons)
Added Clipboard support, which works if you don't lose characters :P

Oh and corrected the name and bumped the version