Colors!
Default color is now light gray :< but that's how it was on JTerminal
anyway (that i can remember)

Sequence m is implemented! (atleast for stuff possible on a OC Screen)
extended opcodes -4 and -5 (set BG and FG) to be generic FG and BG
setters
opcode -1004 (set cursor character) will now restore the current FG and
BG setting when turning off the cursor.

Oh and i had to add all of this to the persistence stuff